+++ src/client/js/components/Admin/Users/SendInvitationMailButton.jsx
